backend/requirements.txt
django==3.0.4
psycopg2==2.7.5
djangorestframework==3.11.0